ping_pong/requirements.txt
pygame==2.4.0
Pillow~=9.5.0